Chore:(access control for guests and users)

+++ app/Http/Controllers/PostsController.php
@@ -14,7 +14,7 @@ class PostsController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth');
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
     /**
      * Display a listing of the resource.
@@ -86,6 +86,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Status error 401 | Unauthorized Page ');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -121,6 +124,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Status error 401 | Unauthorized Page ');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post successfully deleted');
     }
